Also dereference the thrown value in Api. Fix the catching in Api; better handling of G__exception (e.g. #include its header first so CINT knows it). Now that we can handle throwing of float, double, void* etc, enable the tests that come with it. Changes to Api.cxx also fix CINT test t977.
git-svn-id: http://root.cern.ch/svn/root/trunk@29450 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/test/t977.h
@@ -89,7 +89,7 @@ void throw_int(void) throw(int) {throw 1234;}
 void throw_long(void) throw(long) {throw 5678L;}
 static int xxx=999;
 void throw_void(void) throw(void*) {throw((void*)(&xxx));}
-void throw_float(void) throw(float) {throw 1.23;}
+void throw_float(void) throw(float) {throw (float)1.23;}
 void throw_double(void) throw(double) {throw 3.14;}
 
 #ifdef __MAKECINT__
